requirements.txt
tensorflow>=2.15.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
streamlit>=1.30.0
numpy>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
